Fix players.rating column type: int -> numeric
tt.league-pro.com reports ratings with decimals (e.g. 472.4, 350.39),
but the column was int, so every player upsert in sync.js failed —
which is why the last run showed matchesProcessed: 5, picksGenerated: 0
with no visible cause until the error-logging fix surfaced it.

+++ supabase/schema.sql
@@ -5,7 +5,7 @@
 create table if not exists players (
   id           bigint primary key,          -- id del jugador en tt.league-pro.com
   name         text not null,
-  rating       int,
+  rating       numeric,
   league_range text,                        -- ej. "800-900"
   updated_at   timestamptz default now()
 );
